groups all token-related liftables in one file
Apparently this is enough to satisfy materializeAdt.
